調整帳號資訊頁結構，將稅務資訊、持股資訊、玩家紀錄， 參照石頭資訊和持有產品的模式，獨立出來
