+++ client/accountInfo/accountInfoStonePanel.js
@@ -1,7 +1,6 @@
 import { Meteor } from 'meteor/meteor';
 import { Template } from 'meteor/templating';
 import { ReactiveVar } from 'meteor/reactive-var';
-import { FlowRouter } from 'meteor/kadira:flow-router';
 
 import { dbCompanyStones, stoneTypeList, stonePowerTable } from '/db/dbCompanyStones';
 import { inheritedShowLoadingOnSubscribing } from '../layout/loading';
@@ -15,7 +14,7 @@ Template.accountInfoStonePanel.onCreated(function() {
   this.placedStonesOffset = new ReactiveVar(0);
 
   this.autorunWithIdleSupport(() => {
-    const userId = FlowRouter.getParam('userId');
+    const { user: { _id: userId } } = Template.instance().data;
 
     if (userId) {
       const offset = this.placedStonesOffset.get();
