+++ client/accountInfo/accountLogViewer.html
@@ -1,0 +1,31 @@
+<template name="accountLogViewer">
+  {{#if onlyViewAccuse}}
+    {{> accountAccuseLogList}}
+  {{else}}
+    {{> accountInfoLogList}}
+  {{/if}}
+</template>
+
+<template name="accountAccuseLogList">
+  <button class="btn btn-outline-info btn-sm mt-1">檢視所有紀錄</button>
+  {{#each accuseList}}
+    {{> displayLog}}
+  {{else}}
+    查無資料！
+  {{/each}}
+  {{#with paginationData}}
+    {{> pagination}}
+  {{/with}}
+</template>
+
+<template name="accountInfoLogList">
+  <button class="btn btn-outline-info btn-sm mt-1">只檢視金管會相關紀錄</button>
+  {{#each logList}}
+    {{> displayLog}}
+  {{else}}
+    查無資料！
+  {{/each}}
+  {{#with paginationData}}
+    {{> pagination}}
+  {{/with}}
+</template>
